asdf

+++ bronze/simulation/cses-subarraysums2/sums.cpp
@@ -15,9 +15,9 @@ int main() {
         cin >> arr[i];
     }
 
-    unordered_map<int, int> sums;
-    int prefix = 0;
-    int ans = 0;
+    unordered_map<long long, int> sums;
+    long long prefix = 0;
+    long long ans = 0;
     sums[0] = 1;
     for (int i = 1; i <= n; i++) {
         prefix += arr[i];
